Optimizing models and preprocessing
